requirements.txt
# python3 -m pip install -r requirements.txt --force-reinstall
html5lib==1.0b8
lxml==4.9.2
beautifulsoup4>=4.4.1
html-table-parser-python3>=0.1.0
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
Requests==2.30.0
seaborn==0.12.2